feat: add authentication and authorization using sso and pls
* feat: initial test for oidc

* feat: working basic authentication

* feat: add authorization

* fix: add error handling and env variables

* fix: error handle missing .env

+++ backend/src/supplier.rs
@@ -1,6 +1,10 @@
+use actix_identity::Identity;
+use actix_session::Session;
 use actix_web::{delete, get, patch, post, web, HttpResponse, Responder};
 use serde::{Deserialize, Serialize};
 use sqlx::{Pool, Sqlite};
+
+use crate::auth::check_auth;
 
 #[derive(Debug, Serialize, Deserialize)]
 struct SupplierGetResponse {
@@ -40,11 +44,19 @@ struct Query {
 #[get("/{club}/supplier")]
 pub(crate) async fn get_supplier(
     club: web::Path<String>,
+    id: Option<Identity>,
+    session: Session,
     pool: web::Data<Pool<Sqlite>>,
     query: web::Query<Query>,
 ) -> impl Responder {
     log::info!("get supplier");
+
     let club = club.as_ref();
+
+    if !check_auth(id, session, club).await {
+        return HttpResponse::Unauthorized().finish()
+    } 
+
     if let Some(id) = query.id {
         match sqlx::query!("SELECT name FROM suppliers WHERE id = $1", id)
             .fetch_one(pool.get_ref())
@@ -72,16 +84,23 @@ pub(crate) async fn get_supplier(
 pub(crate) async fn add_supplier(
     body: String,
     club: web::Path<String>,
+    id: Option<Identity>,
+    session: Session,
     pool: web::Data<Pool<Sqlite>>,
 ) -> HttpResponse {
     log::info!("add supplier");
     log::debug!("{}", body);
+
+    let club = club.as_ref();
+
+    if !check_auth(id, session, club).await {
+        return HttpResponse::Unauthorized().finish()
+    } 
+
     let supplier: SupplierAddRequest = match serde_json::from_str(&body) {
         Ok(item) => item,
         Err(_) => return HttpResponse::BadRequest().finish(),
     };
-
-    let club = club.as_ref();
 
     match sqlx::query!(
         "INSERT INTO suppliers (name, link, notes, username, password, updated, club) VALUES ($1, $2, $3, $4, $5, strftime('%s', 'now'), $6)",
@@ -104,15 +123,23 @@ pub(crate) async fn add_supplier(
 pub(crate) async fn update_supplier(
     club: web::Path<String>,
     body: String,
+    id: Option<Identity>,
+    session: Session,
     pool: web::Data<Pool<Sqlite>>,
 ) -> impl Responder {
     log::info!("update supplier");
     log::debug!("{}", body);
+
+    let club = club.as_ref();
+
+    if !check_auth(id, session, club).await {
+        return HttpResponse::Unauthorized().finish()
+    } 
+
     let supplier: SupplierUpdateRequest = match serde_json::from_str(&body) {
         Ok(item) => item,
         Err(_) => return HttpResponse::BadRequest().finish(),
     };
-    let club = club.as_ref();
 
     match sqlx::query!(
         "UPDATE suppliers SET name = $1, link = $2, notes = $3, username = $4, password = $5, updated = strftime('%s', 'now') WHERE id = $6 AND club = $7",
@@ -135,14 +162,20 @@ pub(crate) async fn update_supplier(
 #[delete("/{club}/supplier")]
 pub(crate) async fn delete_supplier(
     club: web::Path<String>,
-    id: web::Query<i64>,
+    item_id: web::Query<i64>,
+    id: Option<Identity>,
+    session: Session,
     pool: web::Data<Pool<Sqlite>>,
 ) -> impl Responder {
     let club = club.as_ref();
 
+    if !check_auth(id, session, club).await {
+        return HttpResponse::Unauthorized().finish()
+    } 
+
     match sqlx::query!(
         "DELETE FROM suppliers WHERE id = $1 AND club = $2",
-        id.0,
+        item_id.0,
         club
     )
     .execute(pool.get_ref())
